requirements.txt
Flask
pypdf
Pillow